Crate fedimint_core_server

Modules

client

encode

server

Fedimint Core Server module interface

Macros

module_dyn_newtype_impl_encode_decode

Implement Encodable and Decodable for a “module dyn newtype”

module_plugin_trait_define

Define a “plugin” trait

Structs

ApiEndpoint

Api Endpoint exposed by a server side module

ApiHandler

ApiEndpoint handler exposed by the server side module

ConsensusItem

Input

An owned, immutable input to a Transaction

InputMeta

KeyPair

Opaque data structure that holds a keypair consisting of a secret and a public key.

Output

An owned, immutable output of a Transaction

OutputOutcome

An owned, immutable output of a Transaction before it was finalized

PendingOutput

An owned, immutable output of a Transaction before it was finalized

ServerModule

Signature

SpendableOutput

An owned, immutable output of a Transaction after it was finalized (so it’s spendable)

Transaction

Transaction that was already signed

VerificationCache

Enums

Error

FinalizationError

Traits

IServerModule

Backend side module interface

ModuleApiHandler

ModuleCommon

Common functionality of a Fedimint module

ModuleConsensusItem

ModuleInput

Something that can be an Input in a Transaction

ModuleOutput

Something that can be an Output in a Transaction

ModuleOutputOutcome

ModulePendingOutput

A pending output - tracked and persisted by the client

ModuleSpendableOutput

A spendable output - tracked and persisted by the client

ModuleVerificationCache

PluginConsensusItem

PluginInput

PluginOutput

PluginOutputOutcome

PluginPendingOutput

PluginSpendableOutput

PluginVerificationCache

ServerModulePlugin

Type Definitions

ModuleKey

A module key identifing a module